giovi-ai/email-agent-service/pyproject.toml
[project]
name = "email-agent-service"
version = "0.1.0"
description = "FastAPI service for Gmail ingestion and AI concierge workflows"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "firebase-admin>=6.5.0",
  "google-cloud-firestore>=2.16.0",
  "google-cloud-secret-manager>=2.20.0",
  "google-auth-oauthlib>=1.2.0",
  "google-api-python-client>=2.187.0",
  "google-auth-httplib2>=0.2.1",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "email-validator>=2.2.0",
  "beautifulsoup4>=4.14.2",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "types-python-dateutil",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

